feat(altium): 改进符号引脚导出以支持更丰富的IEEE装饰
- 优先从display控制层获取显示属性，兼容旧showName字段
- 优先从style语义层获取引脚样式，兼容旧hasDot/hasClock字段
- 添加对多种PinDecoration枚举的支持，包括集电极开路、发射极开路、高阻等
- 修正isHidden逻辑以正确处理显示状态

refactor(ir): 提取几何归一化和IR构建工具类

- 创建GeometryNormalizer.h用于坐标解析和几何变换
- 创建IrBuilder.h用于SymbolData到SymbolComponentIR的转换
- 从FootprintDataConverter.h中移除对SymbolDataConverter.h的依赖
- 将坐标解析工具函数从SymbolDataConverter分离到GeometryNormalizer
- 添加丰富的PinDecoration枚举定义支持更多IEEE图形装饰
